corrected database

+++ ManufacturingCompany/App_Data/tblPayroll.sql
@@ -3,7 +3,7 @@
     [employee_id]            NVARCHAR (128) NOT NULL,
     [period_begin]           DATETIME       NOT NULL,
     [period_end]             DATETIME       NOT NULL,
-    [total_hours]            FLOAT NULL,
+    [total_hours]            FLOAT (53)     NULL,
     [subtotal]               MONEY          NOT NULL,
     [federal_tax_witholding] MONEY          NOT NULL,
     [state_tax_witholding]   MONEY          NOT NULL,
